# performance: date-time rules in collect module