With large libraries, the 'date-time' rule in the collect module becomes very slow, as there is basically one entry in the tree view per image in the library.  This not only slows down switching collections, it also slows down any user actions which might potentially change the images contained in the collection.

For a library containing 155k images, the times of the various phases of the tree_view() function are:
-   0.35 sec   SQL lookup (~lines 1345-1380)
-   0.03 sec   sort (~line 1385)
-   1.67 sec   build treeview (~lines 1390-1520)
-   0.41 sec   expand treeview entries corresponding to selected range (~1550-1600)

This patch turns off the tree view's automatic sorting (since elements are added in the desired order), uses prepend instead of append for better efficience of insertion, and `gtk_tree_store_insert_with_values` instead of `gtk_tree_store_insert`+`gtk_tree_store_set` for reduced Gtk signalling.  That reduces the time to build the treeview object from 1.67 seconds to 1.03.

In addition, the string casefolding in the SQL lookup loop is moved inside the only branch where it is actually used, speeding up that loop from 0.35 seconds to 0.19 seconds for date-time rules.

Finally, half (on average) of the 0.41 second to expand the branches of the treeview are saved by stopping iteration on reaching the target branch.

**Additional notes**
- It should be possible to shave an additional 0.4 seconds off the construction time with a more efficient method of propagating counts upward -- disabling that code entirely cuts 0.5 second -- but that's a more involved change than the ones this PR makes.
- The timeline code should use a 'date' rule instead of 'date-time' when possible; a 'date' rule on 155k images takes 0.07 seconds for the SQL lookup, under 0.01 for sorting, 0.01 for construction, and under 0.01 seconds for expansion due to the far smaller number of treeview entries.